some unit tests for our elusive display value
